+++ frontend/src/components/common/Header/HomePageHeader/HomePageHeader.tsx
